Add OTLP-119 conformance test: HTTP client semantic convention validation
Implements comprehensive test validating that spans with kind=CLIENT and
instrumentation.scope.name="http.client" must have http.url or url.full
attribute set per OTLP semantic conventions.

+++ src/observability/otel_structured_concurrency.rs
@@ -69,7 +69,7 @@ pub enum EntityId {
 impl EntityId {
     /// Creates a region entity ID from a raw number for testing.
     pub fn region_from_raw(id: u64) -> Self {
-        Self::Region(RegionId::new_for_test(id, 1))
+        Self::Region(RegionId::new_for_test(id as u32, 1))
     }
 }
